add chart with weather conditions from last 24h

+++ src/main/java/com/example/googlechartsthymeleaf/entity/outside_weather/CurrentWeatherEntity.java
@@ -43,10 +43,10 @@ public class CurrentWeatherEntity {
     private Float windSpeed;
     private Short windDegree;
     private Float windGustSpeed;
-    private Integer measurementTime;
+    private Long measurementTime;
     private String country;
-    private Integer sunriseTime;
-    private Integer sunsetTime;
+    private Long sunriseTime;
+    private Long sunsetTime;
     private Short timezoneOffset;
     private String cityName;
